refactor: 디렉토리 구조 개선 - shared 추가하여 공통 모듈의 응집도 강화
