+++ src/pages/createTask.jsx
@@ -2,10 +2,10 @@ import React, { useState, useMemo } from 'react'
 import { Link, useNavigate, useLocation } from 'react-router-dom'
 
 import styled from 'styled-components'
-import NavBar from '../components/NavBar'
-import Text from '../components/Text'
-import Button from '../components/Button'
-import Input from '../components/Input'
+import NavBar from '../shared/components/NavBar'
+import Text from '../shared/components/Text'
+import Button from '../shared/components/Button'
+import Input from '../shared/components/Input'
 
 const BUTTON_TEXT = Object.freeze({
 	VALID: '다음 단계',
