+++ src/shared/components/NavBar.jsx
